// HelpDeskFE/HelpDeskFETp/src/api/model/createSupplierMeetingDto.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * HelpDesk API
 * OpenAPI spec version: v1
 */

export interface CreateSupplierMeetingDto {
  idSupplier: number;
  /**
     * @maxLength 500
     * @nullable
     */
  description?: string | null;
  /**
     * @minLength 1
     * @maxLength 500
     */
  reasonWork: string;
  /** @nullable */
  evidence?: string | null;
  meetingDate: string;
}